requirements.txt
python-dotenv
fastapi
uvicorn
psycopg2-binary
dbt-core
dbt-postgres
telethon
# Install CPU-only PyTorch and related libraries first
torch --index-url https://download.pytorch.org/whl/cpu
torchvision --index-url https://download.pytorch.org/whl/cpu
torchaudio --index-url https://download.pytorch.org/whl/cpu
ultralytics
dagster
dagster-webserver
